+++ druid/base/drag.lua
@@ -172,7 +172,7 @@ end
 
 function M.on_input(self, action_id, action)
 	if action_id ~= const.ACTION_TOUCH and action_id ~= const.ACTION_MULTITOUCH then
-		return
+		return false
 	end
 
 	if not helper.is_enabled(self.node) then